feat: wrap RichText into component
